rename resource_evaluator_val_loop
