+++ mmrazor/engine/runner/__init__.py
@@ -2,8 +2,8 @@
 from .autoslim_val_loop import AutoSlimValLoop
 from .darts_loop import DartsEpochBasedTrainLoop, DartsIterBasedTrainLoop
 from .distill_val_loop import SingleTeacherDistillValLoop
-from .evaluator_val_loop import ResourceEvaluatorLoop
 from .evolution_search_loop import EvolutionSearchLoop
+from .resource_evaluator_val_loop import ResourceEvaluatorLoop
 from .slimmable_val_loop import SlimmableValLoop
 from .subnet_sampler_loop import GreedySamplerTrainLoop
